# improve performance of binary morphology for fully nonzero structuring elements